+++ ANTWebAPI/src/ANTWebAPI.DAL/Repositories/CatalogRepository.cs
@@ -5,14 +5,15 @@ using Microsoft.EntityFrameworkCore;
 
 namespace ANTWebAPI.DAL.Repositories;
 
-internal class CatalogRepository(ANTDbContext context) : IRepository<Catalog>
+public class CatalogRepository(ANTDbContext context) : IRepository<Catalog>
 {
     private readonly ANTDbContext _context = context;
 
     public async Task<bool> DeleteAsync(long id)
     {
         if (await _context.Articles.AnyAsync(e => e.CatalogId == id)) return false;
-        var model = await _context.Catalogs.FirstAsync(e => e.Id == id);
+        var model = await _context.Catalogs.FirstOrDefaultAsync(e => e.Id == id);
+        if (model == null) return false;
         _context.Catalogs.Remove(model);
         await _context.SaveChangesAsync();
         return true;
